Fixed property loading caused by improper specs.
Well then.  This was interesting - the file format specifications are
weird sometimes.  Let me quote the specifications for the Map Export
Poperties record.


   Name   |   Size   |      Content      
----------|----------|-------------------
Property  | Integer  | Selected property
Size      | Integer  | Selected size

Ok, now let me quote the code used by the origional app to load it.

    const ProGBOMapExportPropID = $0A;
      type
        RProGBOMapExportPropDefRec = record
          Nr     : byte;
          Size   : integer;
      end;
    type TProGBOMapExportProp = packed record
      Props       : array[0..0] of RProGBOMapExportPropDefRec;  (*
PropCnt *)
    end;
    PProGBOMapExportProp = ^TProGBOMapExportProp;

Now, do you notice something?  Let's see.  Nr is a byte.  And guess
what?  It's an int in the format specifications.

It appears that it is padded to the size of an int, however.  Which
actually makes things worse - *most* of the time it pads with 0's.  But
it seems that ocasionally it doesn't, but instead uses whatever data was
there before, resulting in garbage data.

My fix is to simply only use the first byte of the loaded data.  It's
kinda weird and ugly, but it works... well enough.